requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.1
MarkupSafe==2.1.4
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==1.10.13
pydantic-sqlalchemy==0.0.9
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.4
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
